requirements.txt
acme==1.1.0
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
autobahn==23.1.2
cachetools==5.3.0
channels==3.0.5
channels-redis==3.3.1
chardet==3.0.4
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==39.0.1
daphne==3.0.2
distlib==0.3.6
distro==1.4.0
Django==3.2.16
django-cors-headers==3.11.0
django-redis==5.2.0
djangorestframework==3.12.4
filelock==3.9.0
future==0.18.3
hiredis==2.2.1
hyperlink==21.0.0
idna==2.8
# isc==2.0
josepy==1.2.0
mock==3.0.5
msgpack==1.0.4
mysqlclient==2.1.1
netifaces==0.10.4
oauth2==1.9.0.post1
parsedatetime==2.4
pbr==5.4.5
Pillow==9.2.0
platformdirs==3.0.0
ply==3.11
protobuf==4.21.12
pycparser==2.21
pyCryptox==3.18.0
PyICU==2.4.2
pyOpenSSL==23.0.0
pyRFC3339==1.1
pytz==2022.7.1
PyYAML==5.3.1
redis==4.5.1
requests==2.22.0
requests-toolbelt==0.8.0
rsa==4.9
six==1.14.0
sqlparse==0.4.3
txaio==23.1.1
urllib3==1.25.8
zope.component==4.3.0
zope.event==4.4
# zope.hookable==5.0.0
# zope.interface==4.7.1

# local server requirements

# # pip install python-dateutil
# # pip install pyyaml ua-parser user-agents
# # pip install django-user-agents
# acme==1.1.0
# aioredis==1.3.1
# asgiref==3.6.0
# async-timeout==4.0.2
# autobahn==23.1.2
# cachetools==5.3.0
# channels==3.0.5
# channels-redis==3.3.1
# chardet==3.0.4
# ConfigArgParse==0.13.0
# configobj==5.0.6
# cryptography==39.0.1
# daphne==3.0.2
# distlib==0.3.6
# distro==1.4.0
# Django==3.2.16
# django-cors-headers==3.11.0
# django-redis==5.2.0
# djangorestframework==3.12.4
# filelock==3.9.0
# future==0.18.3
# hiredis==2.2.1
# hyperlink==21.0.0
# idna==2.8
# # isc==2.0
# josepy==1.2.0
# mock==3.0.5
# msgpack==1.0.4
# netifaces==0.10.4
# oauth2==1.9.0.post1
# parsedatetime==2.4
# pbr==5.4.5
# Pillow==9.2.0
# platformdirs==3.0.0
# ply==3.11
# protobuf==4.21.12
# pycparser==2.21
# pyCryptox==3.18.0
# pyOpenSSL==23.0.0
# pyRFC3339==1.1
# pytz==2022.7.1
# PyYAML==5.3.1
# redis==4.5.1
# requests==2.22.0
# requests-toolbelt==0.8.0
# rsa==4.9
# six==1.14.0
# sqlparse==0.4.3
# txaio==23.1.1
# urllib3==1.25.8
# zope.component==4.3.0
# zope.event==4.4
# zope.hookable==5.0.0
# zope.interface==4.7.1


# dev server requirements

# pip install python-dateutil
# pip install pyyaml ua-parser user-agents
# pip install django-user-agents

